Wip on achievement calculation. Changed point data type to float.

+++ db/migrate/20120907103559_create_achievement_points.rb
@@ -3,7 +3,7 @@ class CreateAchievementPoints < ActiveRecord::Migration
     create_table :achievement_points do |t|
       t.integer :user_id, :required => true, :null => false
       t.integer :achievement_id, :required => true, :null => false
-      t.integer :point, :required => true, :null => false
+      t.float :point, :required => true, :null => false
       t.string  :description
       t.timestamps
     end
